Apply translate to Arc
