+++ shoes-swt/lib/shoes/swt/arc.rb
@@ -18,7 +18,8 @@ class Shoes
       def initialize(dsl, app)
         @dsl = dsl
         @app = app
-        @painter = Painter.new(self)
+
+        @painter = ArcPainter.new(self)
         @app.add_paint_listener @painter
       end
 
@@ -38,33 +39,6 @@ class Shoes
 
       def radians_to_degrees(radians)
         radians * 180 / ::Shoes::PI
-      end
-
-      class Painter < Common::Painter
-        def fill(graphics_context)
-          if @obj.wedge?
-            graphics_context.fill_arc(@obj.element_left, @obj.element_top,
-                                      @obj.element_width, @obj.element_height,
-                                      @obj.angle1, @obj.angle2 * -1)
-          else
-            path = ::Swt::Path.new(::Swt.display)
-            path.add_arc(@obj.element_left, @obj.element_top,
-                         @obj.element_width, @obj.element_height,
-                         @obj.angle1, @obj.angle2 * -1)
-            graphics_context.fill_path(path)
-          end
-        end
-
-        def draw(graphics_context)
-          sw = graphics_context.get_line_width
-          if @obj.element_left && @obj.element_top && @obj.element_width && @obj.element_height
-            graphics_context.draw_arc(@obj.element_left + sw / 2,
-                                      @obj.element_top + sw / 2,
-                                      @obj.element_width - sw,
-                                      @obj.element_height - sw,
-                                      @obj.angle1, @obj.angle2 * -1)
-          end
-        end
       end
     end
   end
